+++ backend/API-Ultimate/src/main/java/org/ultimateam/apiultimate/service/JoueurService.java
@@ -1,6 +1,8 @@
 package org.ultimateam.apiultimate.service;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import org.ultimateam.apiultimate.model.Equipe;
 import org.ultimateam.apiultimate.model.Joueur;
 import org.ultimateam.apiultimate.repository.JoueurRepository;
@@ -59,6 +61,9 @@ public class JoueurService {
      * @param id L'identifiant du joueur à supprimer.
      */
     public void deleteJoueur(Long id) {
+        if (!joueurRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Le joueur n'existe pas");
+        }
         joueurRepository.deleteById(id);
     }
